+++ crates/ryo3/src/submodules.rs
@@ -2,6 +2,19 @@ use pyo3::prelude::*;
 use pyo3::types::PyDict;
 use pyo3::{intern, wrap_pymodule};
 
+#[cfg(feature = "dirs")]
+#[pymodule(gil_used = false, submodule, name = "dirs")]
+pub fn dirs_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    ryo3_dirs::pymod_add(m)?;
+    Ok(())
+}
+#[cfg(feature = "http")]
+#[pymodule(gil_used = false, name = "http")]
+pub fn http(m: &Bound<'_, PyModule>) -> PyResult<()> {
+    ryo3_http::pymod_add(m)?;
+    Ok(())
+}
+
 #[cfg(feature = "jiter")]
 #[pymodule(gil_used = false, name = "JSON")]
 pub fn json(m: &Bound<'_, PyModule>) -> PyResult<()> {
@@ -9,13 +22,11 @@ pub fn json(m: &Bound<'_, PyModule>) -> PyResult<()> {
     Ok(())
 }
 
-#[cfg(feature = "dirs")]
-#[pymodule(gil_used = false, submodule, name = "dirs")]
-pub fn dirs_module(m: &Bound<'_, PyModule>) -> PyResult<()> {
-    ryo3_dirs::pymod_add(m)?;
-    Ok(())
-}
-
+// #[pymodule(gil_used = false, submodule, name = "reqwest")]
+// pub fn reqwest(m: &Bound<'_, PyModule>) -> PyResult<()> {
+//     ryo3_reqwest::pymod_add(m)?;
+//     Ok(())
+// }
 #[cfg(feature = "xxhash")]
 #[pymodule(gil_used = false, submodule, name = "xxhash")]
 pub fn xxhash(m: &Bound<'_, PyModule>) -> PyResult<()> {
@@ -28,8 +39,14 @@ pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
     #[cfg(feature = "dirs")]
     m.add_wrapped(wrap_pymodule!(dirs_module))?;
 
+    #[cfg(feature = "http")]
+    m.add_wrapped(wrap_pymodule!(http))?;
+
     #[cfg(feature = "jiter")]
     m.add_wrapped(wrap_pymodule!(json))?;
+
+    // #[cfg(feature = "reqwest")]
+    // m.add_wrapped(wrap_pymodule!(reqwest))?;
 
     #[cfg(feature = "xxhash")]
     m.add_wrapped(wrap_pymodule!(xxhash))?;
@@ -40,17 +57,32 @@ pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
         .getattr(intern!(py, "modules"))?
         .downcast_into::<PyDict>()?;
 
-    sys_modules.set_item("ry.xxhash", m.getattr("xxhash")?)?;
-    let attr = m.getattr("xxhash")?;
-    attr.setattr("__name__", "ry.xxhash")?;
+    sys_modules.set_item(intern!(py, "ry.dirs"), m.getattr(intern!(py, "dirs"))?)?;
+    let attr = m.getattr(intern!(py, "dirs"))?;
+    attr.setattr(intern!(py, "__name__"), intern!(py, "ry.dirs"))?;
 
-    sys_modules.set_item("ry.JSON", m.getattr("JSON")?)?;
-    let attr = m.getattr("JSON")?;
-    attr.setattr("__name__", "ry.JSON")?;
+    // http
+    sys_modules.set_item(intern!(py, "ry.http"), m.getattr(intern!(py, "http"))?)?;
+    let attr = m.getattr(intern!(py, "http"))?;
+    attr.setattr(intern!(py, "__name__"), intern!(py, "ry.http"))?;
 
-    sys_modules.set_item("ry.dirs", m.getattr("dirs")?)?;
-    let attr = m.getattr("dirs")?;
-    attr.setattr("__name__", "ry.dirs")?;
+    // JSON
+    sys_modules.set_item(intern!(py, "ry.JSON"), m.getattr(intern!(py, "JSON"))?)?;
+    let attr = m.getattr(intern!(py, "JSON"))?;
+    attr.setattr(intern!(py, "__name__"), intern!(py, "ry.JSON"))?;
+
+    // reqwest (TO MOVE)
+    // sys_modules.set_item(
+    //     intern!(py, "ry.reqwest"),
+    //     m.getattr(intern!(py, "reqwest"))?,
+    // )?;
+    // let attr = m.getattr(intern!(py, "reqwest"))?;
+    // attr.setattr(intern!(py, "__name__"), intern!(py, "ry.reqwest"))?;
+
+    // xxhash
+    sys_modules.set_item(intern!(py, "ry.xxhash"), m.getattr(intern!(py, "xxhash"))?)?;
+    let attr = m.getattr(intern!(py, "xxhash"))?;
+    attr.setattr(intern!(py, "__name__"), intern!(py, "ry.xxhash"))?;
 
     Ok(())
 }